make a command for tests in makefile and reformat clients files
