+++ chatSystem/test/client2.txt
@@ -1,0 +1,19 @@
+\ROOMS
+\JOIN Ellen room1
+Hello, everyone
+what is going on
+la la lal a aaaaaa
+...//'';;[]]
+^v^>>>>^v^
+something wrong what ....
+ha ah hah ahahh ahah ahah
+15654683321, what are the number
+how are u?
+\ROOMS
+\WHO
+\WHERE
+\HELP
+\John it's a private message
+\Alan another private message
+now speak to public  
+\LEAVE
