+++ chatSystem/test/client5.txt
@@ -1,0 +1,19 @@
+\ROOMS
+\JOIN Alan room1
+Hello, everyone
+what is going on
+la la lal a
+...//'';;[]] waht
+^v^>>>>^v^
+something wrong what ....
+ha ah hah ahahh ahah ahah
+1565468315511321, what are the number
+how are you?
+\ROOMS
+\WHO
+\WHERE
+\HELP
+\John it's a private message
+\Ellen another private message
+now speak to public 
+\LEAVE
