--- chatSystem/test/testFile.txt
@@ -1,6 +1,0 @@
-\JOIN nickname room
-\ROOMS
-\LEAVE
-\WHO
-\HELP
-\nickname message
